+++ CHANGELOG.md
@@ -25,6 +25,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Updated project links.
 
+### Testing
+
+- Added test cases for `gethash.wrappers.crc32`.
+
 ## [5.9] - 2023-10-30
 
 ### Changed